tests: fix breaking and turn that into new test
